requirements.txt
# ROS2 Home Assistant Dependencies

# Speech Recognition
faster-whisper
silero-vad
sounddevice
numpy

# LLM
anthropic

# TTS
piper-tts
